+++ podcasts/LargeListSummaryCellHeaderView.swift
@@ -27,6 +27,7 @@ class LargeListSummaryCellHeaderView: UIView {
         let label = ThemeableLabel()
         label.font = .font(ofSize: 15, weight: .medium, scalingWith: .title1)
         label.textColor = ThemeColor.primaryText02()
+        label.numberOfLines = 1
         label.adjustsFontForContentSizeCategory = true
         return label
     }()
@@ -35,7 +36,7 @@ class LargeListSummaryCellHeaderView: UIView {
         let label = ThemeableLabel()
         label.textColor = ThemeColor.primaryText01()
         label.font = .font(ofSize: 22, weight: .bold, scalingWith: .title1)
-        label.numberOfLines = 2
+        label.numberOfLines = 1
         label.adjustsFontForContentSizeCategory = true
         return label
     }()
@@ -97,5 +98,25 @@ class LargeListSummaryCellHeaderView: UIView {
 
         horizontalStack.addArrangedSubview(imageView)
         horizontalStack.addArrangedSubview(verticalStack)
+    }
+
+    override func awakeFromNib() {
+        super.awakeFromNib()
+        updateSize()
+    }
+
+    // MARK: - Dynamic Type support
+
+    func updateSize() {
+        topLabel.updateNumberOfLines(regular: 1, accessibility: 2)
+        bottomLabel.updateNumberOfLines(regular: 1, accessibility: 2)
+    }
+
+    override func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
+        super.traitCollectionDidChange(previousTraitCollection)
+
+        if previousTraitCollection?.preferredContentSizeCategory != traitCollection.preferredContentSizeCategory {
+            updateSize()
+        }
     }
 }