+++ podcasts/SingleEpisodeViewController.swift
@@ -8,21 +8,32 @@ class SingleEpisodeViewController: UIViewController {
     private var cancellables = Set<AnyCancellable>()
     private var category: DiscoverCategory?
 
-    @IBOutlet var episodeTitle: ThemeableLabel!
+    @IBOutlet var episodeTitle: ThemeableLabel! {
+        didSet {
+            episodeTitle.font = .font(ofSize: 18, weight: .semibold, scalingWith: .headline)
+        }
+    }
+
     @IBOutlet var podcastTitle: ThemeableLabel! {
         didSet {
             podcastTitle.style = .primaryText02
+            podcastTitle.font = .font(ofSize: 13, weight: .semibold, scalingWith: .footnote)
         }
     }
 
     @IBOutlet var playButton: PlayPauseLabeledButton!
     @IBOutlet var podcastImage: PodcastImageView!
-    @IBOutlet var typeBadgeLabel: UILabel!
+    @IBOutlet var typeBadgeLabel: UILabel! {
+        didSet {
+            typeBadgeLabel.font = .font(ofSize: 13, weight: .semibold, scalingWith: .footnote)
+        }
+    }
 
     @IBOutlet var duration: ThemeableLabel! {
         didSet {
             duration.text = L10n.unknownDuration
             duration.style = .primaryText02
+            duration.font = .font(ofSize: 13, weight: .semibold, scalingWith: .footnote)
         }
     }
 
@@ -105,6 +116,20 @@ class SingleEpisodeViewController: UIViewController {
     @objc func didSelectEpisode(_ sender: Any) {
         viewModel.didSelectEpisode()
     }
+
+    private func updateSize() {
+        episodeTitle.sizeToFit()
+        podcastTitle.sizeToFit()
+        playButton.sizeToFit()
+        view.sizeToFit()
+    }
+
+    override func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
+        super.traitCollectionDidChange(previousTraitCollection)
+        if traitCollection.preferredContentSizeCategory != previousTraitCollection?.preferredContentSizeCategory {
+            updateSize()
+        }
+    }
 }
 
 extension SingleEpisodeViewController: DiscoverSummaryProtocol {
@@ -117,5 +142,6 @@ extension SingleEpisodeViewController: DiscoverSummaryProtocol {
         self.category = category
 
         typeBadgeLabel.text = (item.title ?? L10n.discoverFeaturedEpisode).uppercased()
+        updateSize()
     }
 }