+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 8.8
 -----
 - Show buffering UI on Full Screen Player [#4051](https://github.com/Automattic/pocket-casts-ios/pull/4051)
-
+- Ensure that temporary download files are properly deleted from the system [#4052](https://github.com/Automattic/pocket-casts-ios/pull/4052)
 
 8.7
 -----
@@ -10,6 +10,14 @@
 - Refresh Now Playing info when Bluetooth device connects [#3964](https://github.com/Automattic/pocket-casts-ios/pull/3964) 
 - Fix logout issue when refreshing in background while device is locked [#3983](https://github.com/Automattic/pocket-casts-ios/pull/3983) 
 - Fix Episode Artwork loading when audio asset image differs from feed [#4017](https://github.com/Automattic/pocket-casts-ios/pull/4017)
+- Fix Single episode loading and support for DT [#4063](https://github.com/Automattic/pocket-casts-ios/pull/4063)
+
+8.6.1
+-----
+- Dynamic Type adjustments [#4055](https://github.com/Automattic/pocket-casts-ios/pull/4055)
+- Dynamic Type: Update Discovery Large Cell layout margins
+[#4057](https://github.com/Automattic/pocket-casts-ios/pull/4057)
+- Dynamic Type: Update Sponsored [#4058](https://github.com/Automattic/pocket-casts-ios/pull/4058)
 
 8.6
 -----